Fixed autocomplete bugs and styled
