+++ app/styles/_filters.scss
@@ -117,6 +117,16 @@
 .selected-filter-details {
   margin-left: 35px;
   line-height: 40px;
+
+  button.close {
+    float: none;
+    display: inline-block;
+  }
+}
+
+.selected-filter-details-spaced {
+  margin-top: 1em;
+  margin-bottom: 1em;
 }
 
 .add-new-filter {
