+++ app/templates/components/coding-typeahead.hbs
@@ -1,0 +1,6 @@
+<input
+  class="input-control {{type}} typeahead"
+  placeholder={{placeholder}}
+  value={{displayValue}}
+  onchange={{action 'updateCode'}}
+  onkeyup={{action 'updateCode'}}>
